added possibility to name the user keys (#7)
#### Type of change

- New feature

#### Description

This PR gives `cryptogen` to capability to create named certificates
for users within a peer organization, as it is currently done with the
`peers`.
Before this fix, it was impossible to specify the name of a user,
leading to the production of user crypto material that looked like:
```text
users/
  Admin@org1.example.com/
  User1@org1.example.com/
```
With this additional functionality, if the user specifies a user with a
Name:
```yaml
Users:
  Count: 1
  Specs:
    - Name: testuser
```
it will generate:
```text
users/
  Admin@org1.example.com/
  User1@org1.example.com/
  testuser@org1.example.com/
```
The possibility of giving names to the user credentials is crucial to
successfully handle the current deployment when using `cryptogen`, since
the Ansible scripts use the hostname to identify the path where to
locate the crypto material.

Signed-off-by: pco <pasquale.convertini@ibm.com>